Move Version class to util module
